+++ templates/shared/content-elements/logo/index.js
@@ -1,15 +1,9 @@
 require('./styles.scss');
 
-const contentElement = require('@bsi-cx/design-standard-library-web/content-elements/base/figure/prototype');
+const contentElement = require('@bsi-cx/design-standard-library-web/content-elements/advanced/logo-with-darkmode/prototype');
 
 module.exports = contentElement(
   require('./template.twig'),
   'logo-pm7eOT',
   'Logo',
-  'logo-part-image-1rKWxD',
-  /*'Logo image',*/
-  'Logo Bild',
-  'logo-part-text-ySZsrw',
-  /*'Logo text'*/
-  'Logo Text'
 ).withStyleConfigs();