+++ templates/landingpage/design.js
@@ -16,7 +16,7 @@ module.exports = cx.design
       .withGroupId('header-Eev4oq')
       .withLabel('Header')
       .withContentElements(
-        require('@bsi-cx/design-standard-library-web/content-elements/advanced/logo-with-darkmode'),
+        require('./../shared/content-elements/logo'),
         require('./../shared/content-elements/header-image')),
     cx.contentElementGroup
       .withGroupId('layout-EhFpQ7')
@@ -65,13 +65,18 @@ module.exports = cx.design
       .withContentElements(
         require('@bsi-cx/design-standard-library-web/content-elements/form/form-container'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/multi-column'),
+        require('@bsi-cx/design-standard-library-web/content-elements/form/if-then/if-layout'),
+        require('@bsi-cx/design-standard-library-web/content-elements/form/if-then/then-group'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/form-field'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/form-field-tel'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/text-area'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/select'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/checkbox'),
         require('@bsi-cx/design-standard-library-web/content-elements/form/radio'),
-        require('@bsi-cx/design-standard-library-web/content-elements/form/poll')),
+        require('@bsi-cx/design-standard-library-web/content-elements/form/poll'),
+        require('@bsi-cx/design-standard-library-web/content-elements/form/audiorecorder'),
+        // require("@bsi-cx/design-standard-library-web/content-elements/form/regex-input"), !!! set properties.regexPattern first !!!
+      ),
     cx.contentElementGroup
       .withGroupId('footer-lJ2X2V')
       .withLabel('Footer')