+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Hint: You will find the list of compatible BSI versions for each template in the
 ## Next
 * Add git pipeline to build master template
 * Add new if-else-group element
-
+* Reintroduce shared logo element extending `logo-with-darkmode` to restore centralized and brand-specific logo support
 
 ## [1.3.9] - 01.10.2025