+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Hint: You will find the list of compatible BSI versions for each template in the
 
 ## Next
 * Add pollDefaultType property
+* Add git pipeline to build master template
+* Add new if-else-group element
+* Reintroduce shared logo element extending `logo-with-darkmode` to restore centralized and brand-specific logo support
 
 
 ## [1.3.9] - 01.10.2025
@@ -19,7 +22,6 @@ Hint: You will find the list of compatible BSI versions for each template in the
 * Add radioOrientationStyle property
 * Add regexPattern property
 * Change documentation of formTooltipStyle property 
-
 
 ## [1.3.8] - 03.06.2025